fix quest map issue fix quest linking and improve map icon scaling
Main fix was to correct an issue when picking up a quest and the ! would stay.

Added a function to fix quest link insertion for WotLK from the database menu. Updated world map icon scaling to dynamically adjust icon sizes when the map is zoomed, ensuring consistent appearance. Also updated the version in the TOC file to 2.04.